+++ frontend/src/model/chain.ts
@@ -1,4 +1,28 @@
-import { EvmChain } from '../chain/evm/EvmChain';
+export enum EvmChain {
+  Arbitrum = 'Arbitrum',
+  Avalanche = 'Avalanche',
+  Base = 'Base',
+  CronosChain = 'CronosChain',
+  BSC = 'BSC',
+  Blast = 'Blast',
+  Ethereum = 'Ethereum',
+  Optimism = 'Optimism',
+  Polygon = 'Polygon',
+  Zksync = 'Zksync',
+}
+
+export const evmChainIds: Record<EvmChain, number> = {
+  [EvmChain.Arbitrum]: 42161,
+  [EvmChain.Avalanche]: 43114,
+  [EvmChain.Base]: 8453,
+  [EvmChain.CronosChain]: 25,
+  [EvmChain.BSC]: 56,
+  [EvmChain.Blast]: 81457,
+  [EvmChain.Ethereum]: 1,
+  [EvmChain.Optimism]: 10,
+  [EvmChain.Polygon]: 137,
+  [EvmChain.Zksync]: 324,
+};
 
 enum OtherChain {
   THORChain = 'THORChain',
